requirements.txt
instaloader==4.11
python-dotenv==1.0.1
pyinstaller==6.6.0
pandas
openpyxl